[no commit message]

+++ src/java/ff/server/SerialPortCommServer.java
@@ -30,7 +30,7 @@ public class SerialPortCommServer {
     //为巡航，比如削苹果皮等准备。其中Key为ip.value为预置的Y角度，只有达到此角度时才接受其它命令。
     private static Map<String, Integer> isCruisingPresetAngleY = new ConcurrentHashMap<String, Integer>();
     //为了修正在回到断点云台不到位的情况。加上命令调整时间，如果超过1秒，就再发送一次，相同的命令。值为毫秒数。
-    private static Map<String, Integer> breakPointReturnBeginTime = new ConcurrentHashMap<String, Integer>();
+    private static Map<String, Long> breakPointReturnBeginTime = new ConcurrentHashMap<String, Long>();
     //巡航断点，用来记录巡航人为停止时的XY角度。Key为IP,Value为（X|Y）记录二个角度。
     private static Map<String, String> cruiseBreakpoint = new ConcurrentHashMap<String, String>();
     //标记正在为回到断点，继续巡航的云台，提供回到断点状态位的标志。如果正在回到断点则不再发送调整命令。    
@@ -375,7 +375,7 @@ public class SerialPortCommServer {
         return isAdjustingXYForBreakpoint;
     }
 
-    public Map<String, Integer> getBreakPointReturnBeginTime() {
+    public Map<String, Long> getBreakPointReturnBeginTime() {
         return breakPointReturnBeginTime;
     }
 }
